more polishing and introduction of a couple of wrapper types
Some more polishing of the basic code I provided last week.
